requirements.txt
beautifulsoup4==4.6.0
Django==1.9
django-tinymce==2.3.0
filelock==2.0.11
futures==3.0.4
guppy==0.1.10
nltk==3.2.4
oauthlib==1.0.3
PyJWT==1.4.0
pymediawiki==0.3.14
python-openid==2.2.5
python-social-auth==0.2.14
pytz==2017.2
regex==2017.7.28
requests==2.9.1
requests-oauthlib==0.6.1
six==1.10.0
typing==3.6.2
wcwidth==0.1.7
wikipedia==1.4.0
wikitextparser==0.17.1
